Rename quantified pattern file
