+++ src/pattern/quantified-patterns/QuantifiedPath.ts
